// src/app/shared/enums/sweetalert.enum.ts
export enum SweetAlert {
    btnConfirmationColor = '#009CDD',
    btnCancelColor = '#6c757d',
    btnConfirmationText = 'Aceptar',
    btnCancelText = 'Cancelar',
    btnDeleteText = 'Eliminar',
    iconSuccess = 'success',
    iconInfo = 'info',
    iconError = 'error',
    iconWarning = 'warning',
    iconWarningToast = 'warn',
  
    titleAlert = 'Hecho',
    titleAlertError = 'Error',
    titleAlertWarning = 'Cuidado',
    titleAlertRetry = 'Reintento de conexión',
    titleAlertValidationParameters = 'Valor Parámetro',    

    messageAlertCreate = '¡Registro creado satisfactoriamente!',
    messageAlertUpdate = '¡Registro actualizado satisfactoriamente!',
    messageAlertDelete = '¡Registro eliminado satisfactoriamente!',
    messageAlertError = '¡No se pudo ejecutar el proceso!',

  }  